Format and rename files
